extracted Editor component
