+++ packages/playground/src/app.jsx
@@ -1,5 +1,5 @@
-import { useState, createRef, Component } from 'react';
-import MonacoEditor from '@monaco-editor/react';
+import { createRef, Component } from 'react';
+
 import { samples } from './samples';
 import 'react-app-polyfill/ie11';
 import Form, { withTheme } from '@rjsf/core';
@@ -17,6 +17,7 @@ import {
   ValidatorSelector,
   SubthemeSelector,
   RawValidatorTest,
+  Editor,
 } from './components';
 
 const log = (type) => console.log.bind(console, type);
@@ -40,62 +41,6 @@ const liveSettingsSchema = {
     },
   },
 };
-
-const monacoEditorOptions = {
-  minimap: {
-    enabled: false,
-  },
-  automaticLayout: true,
-};
-
-class Editor extends Component {
-  constructor(props) {
-    super(props);
-    this.state = { valid: true, code: props.code };
-  }
-
-  UNSAFE_componentWillReceiveProps(props) {
-    this.setState({ valid: true, code: props.code });
-  }
-
-  shouldComponentUpdate(nextProps, nextState) {
-    if (this.state.valid) {
-      return JSON.stringify(JSON.parse(nextProps.code)) !== JSON.stringify(JSON.parse(this.state.code));
-    }
-    return false;
-  }
-
-  onCodeChange = (code) => {
-    try {
-      const parsedCode = JSON.parse(code);
-      this.setState({ valid: true, code }, () => this.props.onChange(parsedCode));
-    } catch (err) {
-      this.setState({ valid: false, code });
-    }
-  };
-
-  render() {
-    const { title } = this.props;
-    const icon = this.state.valid ? 'ok' : 'remove';
-    const cls = this.state.valid ? 'valid' : 'invalid';
-    return (
-      <div className='panel panel-default'>
-        <div className='panel-heading'>
-          <span className={`${cls} glyphicon glyphicon-${icon}`} />
-          {' ' + title}
-        </div>
-        <MonacoEditor
-          language='json'
-          value={this.state.code}
-          theme='vs-light'
-          onChange={this.onCodeChange}
-          height={400}
-          options={monacoEditorOptions}
-        />
-      </div>
-    );
-  }
-}
 
 class Playground extends Component {
   constructor(props) {
