+++ packages/playground/src/components/index.ts
@@ -5,3 +5,4 @@ export { Selector } from './Selector';
 export { ValidatorSelector } from './ValidatorSelector';
 export { SubthemeSelector } from './SubthemeSelector';
 export { RawValidatorTest } from './RawValidatorTest';
+export { Editor } from './Editor';
